feat: use own loader for modules
